Provide a disconnect only option that is safe to use with transactions
Transparent retry as this gem currently performs it is not safe when using transactions can introduce data
integrity issues.

This commit adds an option to simply disconnect on
read_only errors. Leaving the application to deal with the failover
event as it would a normal disconnection.

This allows transactions to be
restarted as needed and will not leave the application in a dangerous
state where it's operating in autocommit mode, but belives it's inside a
transaction.